Add Renderer to scripting

+++ src/input.h
@@ -19,9 +19,9 @@ class Input {
         virtual bool isHeld(size_t key) const;
         virtual bool isPressed(size_t key) const;
         virtual bool isReleased(size_t key) const;
-        virtual bool mouseHeld(uint8_t button) const;
-        virtual bool mousePressed(uint8_t button) const;
-        virtual bool mouseReleased(uint8_t button) const;
+        virtual bool mouseHeld(size_t button) const;
+        virtual bool mousePressed(size_t button) const;
+        virtual bool mouseReleased(size_t button) const;
         virtual Vec mousePos() const;
 };
